Fix SFTP: Test Connection requiring sftp.json to be the active editor
commandTestConnection resolved the config by stripping two segments off the
active editor's path and matching that against service.workspace, so it only
ever worked when .vscode/sftp.json itself was focused. Running the command
from the Command Palette while editing any other file failed with "Could not
find a matching SFTP config for this file."

Promote the resolution commandToggleUploadOnSave already had into a shared
resolveTargetService(): explicit uri (context menu / CodeLens) -> the active
editor's owning service via getFileService() -> the workspace's only config
-> QuickPick when several are configured. Both commands now use it.
