+++ src/commands/commandTestConnection.ts
@@ -1,33 +1,17 @@
-import * as path from 'path';
-import { Uri, window } from 'vscode';
+import { Uri } from 'vscode';
 import { COMMAND_TEST_CONNECTION } from '../constants';
 import { showInformationMessage, showErrorMessage } from '../host';
 import logger from '../logger';
-import { getAllFileService } from '../modules/serviceManager';
 import { checkCommand } from './abstract/createCommand';
-
-function normalizePath(p: string) {
-  return path.normalize(p);
-}
-
-function findFileServiceForConfigUri(uri: Uri) {
-  const configDir = normalizePath(path.dirname(path.dirname(uri.fsPath)));
-  return getAllFileService().find(service => normalizePath(service.workspace) === configDir);
-}
+import { resolveTargetService } from './shared';
 
 export default checkCommand({
   id: COMMAND_TEST_CONNECTION,
 
   async handleCommand(uri?: Uri) {
-    const targetUri = uri || (window.activeTextEditor && window.activeTextEditor.document.uri);
-    if (!targetUri) {
-      showErrorMessage('No sftp.json file found to test.');
-      return;
-    }
-
-    const service = findFileServiceForConfigUri(targetUri);
+    const service = await resolveTargetService(uri, 'Select a config to test');
     if (!service) {
-      showErrorMessage('Could not find a matching SFTP config for this file.');
+      showErrorMessage('No SFTP config found to test.');
       return;
     }
